Search: extract connect
